+++ src/main/java/kitchenpos/ordertable/infrastructure/JpaOrderTableRepository.java
@@ -1,7 +1,7 @@
-package kitchenpos.table.infrastructure;
+package kitchenpos.ordertable.infrastructure;
 
 import java.util.List;
-import kitchenpos.table.domain.OrderTable;
+import kitchenpos.ordertable.domain.OrderTable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface JpaOrderTableRepository extends JpaRepository<OrderTable, Long> {
