+++ src/main/java/kitchenpos/ordertable/ui/OrderTableRestController.java
@@ -1,9 +1,9 @@
-package kitchenpos.table.ui;
+package kitchenpos.ordertable.ui;
 
 import java.net.URI;
 import java.util.List;
-import kitchenpos.table.application.TableService;
-import kitchenpos.table.domain.OrderTable;
+import kitchenpos.ordertable.application.OrderTableService;
+import kitchenpos.ordertable.domain.OrderTable;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -13,17 +13,17 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
-public class TableRestController {
+public class OrderTableRestController {
 
-    private final TableService tableService;
+    private final OrderTableService orderTableService;
 
-    public TableRestController(final TableService tableService) {
-        this.tableService = tableService;
+    public OrderTableRestController(final OrderTableService orderTableService) {
+        this.orderTableService = orderTableService;
     }
 
     @PostMapping("/api/tables")
     public ResponseEntity<OrderTable> create(@RequestBody final OrderTable orderTable) {
-        final OrderTable created = tableService.create(orderTable);
+        final OrderTable created = orderTableService.create(orderTable);
         final URI uri = URI.create("/api/tables/" + created.getId());
         return ResponseEntity.created(uri)
             .body(created)
@@ -33,7 +33,7 @@ public class TableRestController {
     @GetMapping("/api/tables")
     public ResponseEntity<List<OrderTable>> list() {
         return ResponseEntity.ok()
-            .body(tableService.list())
+            .body(orderTableService.list())
             ;
     }
 
@@ -43,7 +43,7 @@ public class TableRestController {
         @RequestBody final OrderTable orderTable
     ) {
         return ResponseEntity.ok()
-            .body(tableService.changeEmpty(orderTableId, orderTable))
+            .body(orderTableService.changeEmpty(orderTableId, orderTable))
             ;
     }
 
@@ -53,7 +53,7 @@ public class TableRestController {
         @RequestBody final OrderTable orderTable
     ) {
         return ResponseEntity.ok()
-            .body(tableService.changeNumberOfGuests(orderTableId, orderTable))
+            .body(orderTableService.changeNumberOfGuests(orderTableId, orderTable))
             ;
     }
 }
