+++ src/main/java/kitchenpos/tablegroup/infrastructure/JpaTableGroupDao.java
@@ -1,9 +1,9 @@
-package kitchenpos.table.infrastructure;
+package kitchenpos.tablegroup.infrastructure;
 
 import java.util.List;
 import java.util.Optional;
-import kitchenpos.table.domain.TableGroup;
-import kitchenpos.table.domain.TableGroupDao;
+import kitchenpos.tablegroup.domain.TableGroup;
+import kitchenpos.tablegroup.domain.TableGroupDao;
 import org.springframework.context.annotation.Primary;
 import org.springframework.stereotype.Repository;
