+++ src/test/java/kitchenpos/dao/OrderDaoTest.java
@@ -7,8 +7,8 @@ import java.util.Arrays;
 import java.util.List;
 import kitchenpos.order.domain.Order;
 import kitchenpos.order.domain.OrderStatus;
-import kitchenpos.table.domain.OrderTableDao;
-import kitchenpos.table.domain.OrderTable;
+import kitchenpos.ordertable.domain.OrderTableDao;
+import kitchenpos.ordertable.domain.OrderTable;
 import kitchenpos.order.domain.OrderDao;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
