+++ src/test/java/kitchenpos/dao/OrderLineItemTest.java
@@ -9,8 +9,8 @@ import kitchenpos.menu.domain.MenuDao;
 import kitchenpos.order.domain.Order;
 import kitchenpos.order.domain.OrderLineItem;
 import kitchenpos.order.domain.OrderStatus;
-import kitchenpos.table.domain.OrderTableDao;
-import kitchenpos.table.domain.OrderTable;
+import kitchenpos.ordertable.domain.OrderTableDao;
+import kitchenpos.ordertable.domain.OrderTable;
 import kitchenpos.order.domain.OrderDao;
 import kitchenpos.order.domain.OrderLineItemDao;
 import org.junit.jupiter.api.Test;
