+++ src/test/java/kitchenpos/dao/TableGroupDaoTest.java
@@ -3,8 +3,8 @@ package kitchenpos.dao;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.time.LocalDateTime;
-import kitchenpos.table.domain.TableGroup;
-import kitchenpos.table.domain.TableGroupDao;
+import kitchenpos.tablegroup.domain.TableGroup;
+import kitchenpos.tablegroup.domain.TableGroupDao;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
