+++ src/test/java/kitchenpos/integration/FixtureMaker.java
@@ -10,18 +10,18 @@ import kitchenpos.menugroup.domain.MenuGroupDao;
 import kitchenpos.menu.domain.MenuProductDao;
 import kitchenpos.order.domain.OrderDao;
 import kitchenpos.order.domain.OrderLineItemDao;
-import kitchenpos.table.domain.OrderTableDao;
+import kitchenpos.ordertable.domain.OrderTableDao;
 import kitchenpos.product.domain.ProductDao;
-import kitchenpos.table.domain.TableGroupDao;
+import kitchenpos.tablegroup.domain.TableGroupDao;
 import kitchenpos.menu.domain.Menu;
 import kitchenpos.menugroup.domain.MenuGroup;
 import kitchenpos.menu.domain.MenuProduct;
 import kitchenpos.order.domain.Order;
 import kitchenpos.order.domain.OrderLineItem;
 import kitchenpos.order.domain.OrderStatus;
-import kitchenpos.table.domain.OrderTable;
+import kitchenpos.ordertable.domain.OrderTable;
 import kitchenpos.product.domain.Product;
-import kitchenpos.table.domain.TableGroup;
+import kitchenpos.tablegroup.domain.TableGroup;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
