+++ src/test/java/kitchenpos/integration/OrderIntegrationTest.java
@@ -5,8 +5,8 @@ import java.util.List;
 import kitchenpos.order.domain.Order;
 import kitchenpos.order.domain.OrderLineItem;
 import kitchenpos.order.domain.OrderStatus;
-import kitchenpos.table.domain.OrderTable;
-import kitchenpos.table.domain.TableGroup;
+import kitchenpos.ordertable.domain.OrderTable;
+import kitchenpos.tablegroup.domain.TableGroup;
 import kitchenpos.order.dto.OrderCreateRequestDto;
 import kitchenpos.order.dto.OrderCreateResponseDto;
 import kitchenpos.order.dto.OrderUpdateRequestDto;
