+++ src/test/java/kitchenpos/integration/TableIntegrationTest.java
@@ -1,6 +1,6 @@
 package kitchenpos.integration;
 
-import kitchenpos.table.domain.OrderTable;
+import kitchenpos.ordertable.domain.OrderTable;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.MediaType;
